+++ src/Controller/HomeController.php
@@ -6,7 +6,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Mobile_Detect;
-use Proxies\__CG__\App\Entity\Produits;
+use App\Entity\Produits;
 use App\Repository\ProduitsRepository;
 class HomeController extends AbstractController
 {
@@ -28,12 +28,22 @@ class HomeController extends AbstractController
     /**
      * @Route("/produit/{id}", name="show_produit")
      */
-    public function showProduct(int $id, ProduitsRepository $repo) {
+    public function showProduct(Produits $product, ProduitsRepository $repo) {
 
-        $product = $repo->findOneBy(['id' => $id]);
+
+        if ($product->getCategorie() !== null) {
+            $categoryProducts = $repo->createQueryBuilder('p');
+            $categoryProducts->select('p')
+                             ->where('p.categorie = :category')->setParameter('category', $product->getCategorie()->getId())
+                             ->andwhere('p.id != :id')->setParameter('id', $product->getId())
+                             ->setMaxResults(5);
+            $categoryProducts = $categoryProducts->getQuery()->getResult();
+        }
+
 
         return $this->render('produits/product.html.twig', [
-            'product' => $product
+            'product' => $product,
+            'categoryProducts' => (isset($categoryProducts) && !empty($categoryProducts)) ? $categoryProducts : null
         ]);
     }
 }
