+++ templates/back/index.html.twig
@@ -30,7 +30,7 @@
                         <div class="col-sm-3">
                           <div class="card">
                             <div class="card-body d-flex justify-content-between align-items-center">
-                              <h4 class="card-title"><a href="admin_produits" style="color:black;">Produits</a></h4>
+                              <h4 class="card-title"><a href="{{ path('admin_produits')}}" style="color:black;">Produits</a></h4>
                               <a href="{{ path('admin_new_produit') }}" style="font-size:32px;"><i class="fa-solid fa-circle-plus"></i></a>
                             </div>
                           </div>
