+++ templates/produits/allProducts.html.twig
@@ -136,4 +136,5 @@
     {% endblock %}
     {% block javascripts %}
         <script src="{{ asset('js/product/productList.js') }}"></script>
+        <script src="{{ asset('js/cart/cart.js') }}"></script>
     {% endblock %}
