+++ templates/produits/list/productList.html.twig
@@ -2,7 +2,7 @@
     <div class="col-6 col-lg-4 g-mb-30">
         <!-- Product -->
         <figure class="g-pos-rel g-mb-20">
-            <img class="img-fluid" src="{{ asset('uploads/produits/' ~ produit.image1) }}" alt="Image Description">
+            <a href="{{ path('show_produit', {'id': produit.id}) }}"><img class="img-fluid" src="{{ asset('uploads/produits/' ~ produit.image1) }}" alt="Image Description"></a>
 
             <figcaption class="w-100 g-bg-primary g-bg-black--hover text-center g-pos-abs g-bottom-0 g-transition-0_2 g-py-5">
                 <a class="g-color-white g-font-size-11 text-uppercase g-letter-spacing-1 g-text-underline--none--hover" href="">New Arrival</a>
